Fix ahoy dkan test command.
Ref CIVIC-2718

`ahoy dkan test` is failling locally do to ahoy exec/cmd-proxy requiring a new abstraction step.  Some characters now require to be escaped in order to work properly.

The only fix I can figure out for this is to do the step outside of ahoy in a bash script.

AC
==
- [ ] `ahoy dkan test` runs locally.
- [ ] `ahoy dkan test` runs in ci


Former-commit-id: d7d6c78b3dad581170520f732e9e2feea0327d9a